37.3. Ajax Form Response Status Codes
Try/Catch the Ajax Call
Back in our Stimulus controller, now we have the info we need. When the form submit fails validation, the 422 status code will cause the await $.ajax() to throw an exception. So let's wrap this in a try catch block.
Say, try, catch and what we want to do is take out this.modalBodyTarget.innerHTML because we only want to do that on error. In the catch, say this.modalBodyTarget.innerHTML and the response text is available on the error object as e.responseText.
In the successful situation, for now, just console.log('success').
Ok team - let's see what happens! Refresh. Open the modal and let's first submit the form empty. Beautiful! We have errors! Fill in name and price... and submit again. Okay.... nothing happened. But if we check the console... yes! There's our log! All we need to do now is close the modal on success.