37.4. Ajax Form Response Status Codes
Closing the Modal on Success
All we need to do now is close the modal on success.
We can do that by calling hide() on this modal object. The only problem is that... we don't have access to that object from down here in submitForm()! That's ok: this is where having a controller object comes in handy! We can set that modal on a property.
Up at the top of the class... we don't have to do this because JavaScript is friendly, but I'm going to say modal equals null to initialize the property. Then down in connect(), update to this.modal equals new Modal() and this.modal.show().
Now, after the Ajax call, replace console.log('success') with this.modal.hide().
Let's try it now. Refresh - there's my awesome zip drive from the last submit - open the modal, fill out the field and... submit! OMG! We have a fully-functioning Ajax modal system. And... it's completely reusable!
The only imperfect thing is that we don't see the new item on the page, until we refresh. No worries: we'll handle that in the next chapter.