37.5. Ajax Form Response Status Codes
Listening to Bootstrap Modal Events
But before we do, search for "bootstrap 5 modal", click into its docs, and then go down to the "Events" section at the bottom. As the modal opens and closes, the modal itself dispatches some events. What if we needed to listen to those? Like, what if we need to run some custom code whenever a modal is closed. How can we do that? This is the beauty of Stimulus. We already know how! If something dispatches an event, all we need to do is add an action for that event.
Check it out: over in index.html.twig, I'll break this onto multiple lines and then add a data-action="". What we're going to do is listen to this hidden.bs.modal event, which happens after the modal is finished being hidden.
Use that event name, then ->, the name of our controller - modal-form - a # sign and then call the new method modalHidden.
Now, the hidden.bs.modal event won't be dispatched directly on this <div>. It will be dispatched on the modal element. But we already know that's okay! The event will bubble up to this <div>.
Copy modalHidden and head into our Stimulus controller. At the bottom add that method and let's just console.log('it was hidden').
Try it out! Go back to our site, refresh, open the modal and hit cancel. There's the log! Open it again, hit X and... another log. I love that.
Next: to make this a truly smooth user experience, after a successful form submit, what we should really do is make another Ajax call to reload the product list... so that the user can see the new product. Let's do that by, once again, making a reusable Stimulus controller. This controller will be able to reload the HTML for any element.