38.1. Ajax Element Reloading Controller
If we add a new product with valid data - we're going to start selling solar powered flashlights... I got a great deal on them - it does work... but it's not very obvious. You don't see the new product until you reload the page. We could, on success, add a message to the top of the page. That's actually a really great idea. But even if we did, we should also reload the product list so the user can see the new item.
If you think about it, being able to make an Ajax call to a URL... and then use the HTML from that to replace the contents of an element is... kind of a common thing to do. Heck, we basically already do this on the cart page! After removing an item, we make an Ajax call to get the fresh cart list.
So here's the plan: instead of adding more logic to our modal-form controller, which would make it less reusable, let's create a second controller that will make an Ajax call to reload the product list area after a new product is successfully added. To be extra cool - cause we are extra cool - we're going to make this new controller generic so we can reuse it anywhere... like on the cart page!

Bootstrapping the New Controller
Head into templates/product_admin/index.html.twig: the template for the product list page. Let's see. The area that we need to refresh after submit is really just this table. But I'm going to add the new data-controller to the top level div. Let's break it on multiple lines... fix my super old typo on container - no wonder the page didn't look very good - then add {{ stimulus_controller() }}. Call the new controller, how about, reload-content.
Why are we adding it here... and not directly around the table? Well, in order for the reload-controller to know when a form was submitted successfully, we're going to use an old trick: we will dispatch an event from modal-form controller. To listen to that event, reload-controller needs to live on an element that is around both modal-form controller and the <table> that it needs to update.
Let's go add the new file. In assets/controllers/ create reload-content_controller.js. Steal the entire cart-list_controller - since it's so similar - close it and paste. Add a connect() method with console.log() a refresh icon.
Let's give it a go! Refresh the page, check out the logs and... we are connected!