38.2. Ajax Element Reloading Controller
Creating the Partial Endpoint
To be able to refresh the content of this table, we need an endpoint that returns just the table. To do that, we need to isolate the table into a template partial... like we've done before. Copy the entire table, delete it and then, in this same directory - product_admin - create a new file called _list.html.twig. Paste the table here.
Back in index.html.twig include that with {{ include('product_admin/_list.html.twig') }}.
If we refresh now... so far so good: nothing changes.