38.3. Ajax Element Reloading Controller
Like we did with the new product form, the simplest way to create an endpoint that will return just the table is to make the index action capable of returning a full page of HTML - like it's doing right now - or just the _list template partial.
Let's try the same trick as before. First, add a Request argument - the one from HttpFoundation - and then say $template = $request->isXmlHttpRequest(). If this is an Ajax request, use _list.html.twig. Else, use the template we're using now, index.html.twig. Below, we can replace the index.html.twig with $template.
Awesome! Now, copy the route name so we can pass this into our controller. We'll use a value. Start by defining that in our controller. Or, really, just rename cartRefreshUrl to just url and it will be a String. I'll also remove the connect() method.
Pass the value in via the template. So up at the top, on a second argument to stimulus_controller(), set url to path('product_admin_index').
Lovely! Before we use that to make the Ajax call, when that call finishes, we're going to need to know where we should put the new HTML. Let's wrap the table in a new div and make it a target: data-reload-content-target="" and... call it content.
Go set up that target. At the top of the class, add static targets = [] with content inside.
Now let's make the Ajax call. Rename the method to refreshContent(). We're not using this method anywhere... but we will soon. Let's see: we don't need to add this class... the value changed to this.urlValue... and instead of using this.element, use this.contentTarget.
As I mentioned, nobody is calling refreshContent() yet. But if they did, it should make the Ajax call and replace the table with the new HTML. So... how will we call this method?
Well, we need to call it after the modal-form controller finishes submitting successfully. So next: let's dispatch a custom event from modal-form_controller and listen to it so that we can reload the content after it's successful.
Once we're done, we'll prove that our new controller is reusable by completely replacing the cart-list_controller with the new one. Yay for less custom code!