39.2. Dispatching an Event from modal-form
Listening to the Custom Event
Copy that. Now that we've seen the event name, go back over and remove the debug option. Here's the last magic piece to make this work. In index.html.twig, up on the reload-content controller <div>, add an action for the new event: data-action="" the name of the event - modal-form:success - arrow, the name of the controller - reload-content - a pound sign and the name of the method: refreshContent.
That's it! When the modal-form:success event is dispatched, it will bubble up to this element and we will call refreshContent(). Then... that'll take care of the rest!
Let's test it. Reload the page, open the modal and let's sell some avocado peels. Submit. Ah! It... kinda seems like it worked? Except that the Ajax endpoint apparently returned the entire page, not just the template partial.