39.3. Dispatching an Event from modal-form
Using a Query Param instead of isXmlHttpRequest()
Let's go look at the controller. Ah... I warned about this and then did it anyways! We're using fetch() to make the Ajax call... and fetch() does not send the header needed for isXmlHttpRequest() to work. And so, this always renders index.html.twig.
That's okay! Let's just add a query parameter to the end of the URL. I like that better anyways.
Replace this code with $request->query->get('ajax'). So, we'll be looking for a ?ajax=1 on the end of the URL.
In the template - index.html.twig - add that to the URL by passing extra params with ajax set to 1.
Try the form again. Refresh! We'll sell some salsa to go with those avocados... and this time... perfect! The section reloaded. We're done!
If we want to make this a bit fancier, we could even add some classes and use those to force CSS transitions. Or we can do an even simpler trick. When it first starts loading, let's say this.contentTarget.style.opacity = .5.
Copy that, and then, after it finishes loading, set the opacity back to 1.
Add one more product: this time, a mystery box of donuts. Watch the table closely when I hit save. Yes! It was quick, but the table had less transparency for just a moment while it reloaded.
Now that we've got this cool new re-usable reload-content_controller, let's use it to completely replace our custom cart-list controller. Yay for less code! That's next.