40.2. Reusing the "Reload Content" Controller
Making a Target Optional
Head over refresh, remove an item and... uh oh! Let's see:

  Error: missing target element reload-content.content

Ah... In reload-content_controller, we put the HTML into a target called content. We forgot to add that target to the cart page! That's kind of an awesome thing about targets: if you create a controller... and that controller requires a target to be defined, you get a pretty clear error if you forget.
Ok: in cart.html.twig, let's think: which element do we need to reload the content into? Actually it's this <div> right here: the same one that has the data-controller attribute on it. So: we should add a new target to this element, right?
We could. But I have a better idea. What if we make the content target optional?
If it is set, we'll put the HTML into it. But if it is not set, we will assume that the HTML should be put into the top level element: this.element.
How can we do that? Check this out. Say const target = this.hasContentTarget. I mentioned this has thing earlier when we first introduced targets, but we haven't really used it yet. This is a safe way to check whether or not there is a content target defined in the HTML. If there is one, then of course we'll use this.contentTarget. Else use this.element.
Now, down here, use target in those three other places instead of this.contentTarget.
Try it now! Refresh, remove an item and... got it! We get the same functionality with less code!
Next: earlier in the tutorial we built a search-preview controller: oooOOOoo. It works really nicely. But... it would look even better with some CSS fade in and fade out transitions. Is that as easy as just... adding and removing a class at the right time in our controller?
In this case... no. But don't worry: we already have a trick up our sleeve that will allow us to add transitions simply and beautifully.