41.1. CSS Transitions with useTransition -- yarn upgrade stimulus-use@beta
I want to add a CSS transition so that our search preview fades in and out, instead of showing up immediately. This may seem like a small detail, but this kind of stuff can be the difference between a mediocre interface and a truly beautiful experience.

Why CSS Transitions can be Tricky
So how can we do this? We learned some CSS transition basics earlier. In assets/styles/app.css, if you look for cart-item... we set transition: opacity 500ms. When that element also has a removing class, we set opacity to 0.
The result is that, as soon as we added this removing class to an element that already has the cart-item class, that element's opacity changed from 1 to 0 over 500 milliseconds.
So CSS transitions are simple, right?
Well... not always. If we repeated that trick here, we could transition the opacity of this element to 0. But then... the element would still technically be there... it would just be invisible. So if the user clicked right here, to them, it would look like they were clicking this photo. But in reality, they would be clicking the invisible dropdown above, which would take them to the disappearing ink pens page.
Ok then! So we just need to fade this element to opacity 0, wait 500 milliseconds for the transition to finish... and then fully hide the element. Right? Yea! That's exactly what we need to do! But sheesh, this is getting complicated! And if, later, we decided to change the transition in CSS from 500 milliseconds to 1 second, we would need to remember to go into our Stimulus controller and also change the delay there to 1 second... so that we don't hide the element before it's finished fading out.
So that is why CSS transitions are trickier than they seem at first. Fortunately, the stimulus-use library we installed earlier - to get behaviors like clickOutside and dispatch - has a solution for us!
Head to their docs and find a behavior called useTransition. This is a brand new feature that allows us to add CSS transitions when we need to hide or show an element. At the time of recording, this is currently a beta feature... which means it could change a bit - but I'll update the video if that happens.

Upgrading stimulus-use
The feature was introduced in version 0.24.0... and it's so new that I need to upgrade my stimulus-use to get it!

Find your terminal and run:

yarn upgrade stimulus-use@beta
I'm using @beta because the feature isn't included yet in a stable release at the time of this recording. The 0.24.0-1 version is a beta release. Once they release 0.24 stable or higher, you won't need that.
Open up the current search-preview_controller. As a reminder, as we type, this makes an Ajax call to an endpoint that returns the HTML of the search preview, which we then put into resultTarget. That's this element right here.