41.2. CSS Transitions with useTransition
Initializing the Behavior
To activate the useTransition behavior, we'll do the same thing as the other behaviors. Import useTransition... and down in connect(), initialize it: useTransition(this).
But this behavior has several required options. The first is called element. This is the element that we want to hide or show. For us, that's this.resultTarget.
I'm going to paste in the next six options.
These are CSS classes that we're going to create in our CSS file in a minute. Don't worry about them yet. The last option we need is hiddenClass set to d-none.
You'll see how this is used in a second, but d-none is a Bootstrap class that sets display: none on an element.

New enter(), leave() Methods
Before we talk about these options, now that we've initialized the useTransition() behavior, our controller has 3 new methods! Woo! They're called enter(), leave(), and toggle().
Down after we make the Ajax call, to show the resultTarget element, say this.enter().

Then, later, when we want to hide it, we can say this.leave(). And we will not clear the HTML anymore: we'll keep the HTML, but hide the element.

The Elements Hides and Shows!
Okay! That's all we need in the controller. If we do nothing else, the new behavior already gives us an easy way to hide or show an element without CSS transitions.
Let me show you. Reload the page, inspect element on the input and look at the result target element. Woh! It now has a d-none class! useTransition added that!
This happens thanks to the hiddenClass option that we passed. This tells useTransition which class to use to fully hide an element. And, by default, useTransition assumes that your element starts hidden. So as soon as our controller is initialized, it adds that class.
If you have a situation where your element starts visible instead of hidden, you'll need to add a transitioned option set to true... though that may be renamed to an option called initialState that you would set to enter. Check the docs to be sure.
Anyways, watch the element as I start typing. Boom! The d-none class is gone and the element is visible! But there was no transition yet: it all happened instantly. When we click off the element, the d-none is instantly added back. Like I said, if we do nothing else, this behavior gives us an easy way to hide or show an element.
So... how can we make this actually transition? That's up to us! Next, let's add the CSS needed to make the element fade in and fade out. We'll explore the lifecycle of how useTransition intelligently adds and removes classes at just the right time to make this all possible.