44.2. Using the autocomplete-controller
How the autocomplete Ajax Works
Okay. I noticed a few things. To start, by complete chance, the autocomplete controller sends the contents of our input as a query parameter called q... which is exactly what we were using before! You can see that in src/Controller/ProductController.php: we read q as our search term. Awesome!
But we also look for a ?preview query parameter to know if we should render a page partial.
Previously, in search-preview controller, we added that query parameter manually in JavaScript. We can't do that now... but that's ok! We can add it to the url value.
In index.html.twig, back up on the url, add a second argument to path and pass preview: 1. That will fix the full page problem.
But if you try it now... the same thing happened again! An Ajax call was made... but no results are showing up! On the network tab... yeah! It is now returning the partial, not the full page. So... why don't we actually see the results below the input?