45.1. Autocomplete with Transitions
The great thing about the third party stimulus-autocomplete controller is that we were able to quickly create a nice search auto-complete feature with full control over the HTML. And this required zero custom JavaScript.
The only bummer is that we lost our fade-in, fade-out transition! How can we add that back?
When we originally added the transition, we did it by leveraging useTransition() from the stimulus-use library. We added the behavior to the connect() method of a custom controller... then called this.enter() and this.leave() to show and hide the element... with a transition.
But... we can't exactly go into the stimulus-autocomplete source code and hack that stuff in! But we can make this work... as long as this 3rd party controller dispatches the right events.
If you look at their docs... and scroll down: they have an "Events" section! And they do trigger a bunch of events! This toggle one looks perfect:

  Fires when the results element is shown or hidden.

We can use that to trigger our transition.

Using my stimulus-autocomplete Fork
Before we jump in, to get this to work, we need a pull request to be merged into this library, which... at the time of this recording... is still waiting to be merged! Rats! Since it's unmerged, we'll use a fork of this library that contains the tweaks from this pull request. Woo! Living on the edge!
How can we use a fork? Open your package.json file and find the stimulus-autocomplete line. Set its version to a specific branch on my repository: I'll paste that in. This is my username, the name of the library... and then it points to a branch I created called toggle-event-always-dist.
To download this new version, find your terminal and run:

  yarn install