45.2. Autocomplete with Transitions
Creating & Using a 2nd Controller
While that downloads, let's discuss the plan. In order to leverage the useTransition behavior, we're going to need our own custom controller. In assets/controllers/, create a new file called, how about, autocomplete-transition_controller.js. I'll go steal some code from another controller: the only thing we need right now is a connect() method... with console.log('i want transitions').
Back in the template for this page - which lives at templates/product/index.html.twig - we're now going to add two controllers to the same element. We can do that with the stimulus_controller() function... we just need to tweak the format a bit.
The first argument becomes an object... and each controller becomes a key in the object assigned to its values. Once I finish rearranging things... yup: we're passing an object with one controller set to the values for that controller. Now we can add our new controller name - autocomplete-transition - and... this doesn't need any values, so set it to an empty object.
Let's see if the new controller is connected! Find our site... I'll open my dev tools, refresh the page, check the console and... got it! There's our log. If you inspect element on the text box... you can see that the data-controller attribute now has both controllers on it.