45.3. Autocomplete with Transitions
Adding an Action for "toggle"
Ok, let's think: we need our new controller to be notified whenever the autocomplete results should be shown or hidden... so that we can trigger the enter() or leave() transition.
This is where that toggle event comes in handy. The stimulus-autocomplete controller dispatches all of its events on the "main element" that's registered for the controller, which means it's going to be on this div. Let's add an action for the toggle event. Do that with data-action="", the name of the event we want to listen to - toggle - an ->, the name of the controller to call - that's our new custom controller autocomplete-transition - a # sign and the method to call on that controller: let's use toggle.
Copy that and head over to our controller. Add the new toggle() method with an event argument... and console.log(event).
With any luck, whenever the autocomplete results are shown or hidden, we should see this line get hit. Let's see if that happens! Refresh the page. I'll go back to the console and... yes! Actually an event was dispatched immediately when the component initialized. This second one is from when the element was shown. Open up the event object and look at the detail property. Ah: it has a sub-key call action set to the string open! When the results close - like when we click off the area... we see another event. This time action is set to close.
In the toggle() method, we can use this info to either call this.enter() to fade in the element or this.leave() to fade it out.
But in order to even have those methods, we need to initialize the useTransition behavior on this controller.
Let's do that next... but in a way that will allow us to easily reuse our transitions in other controllers. Code-reuse: booya!